+++ src/main/java/twilightforest/client/renderer/entity/RenderTFMinoshroom.java
@@ -42,16 +42,16 @@ public class RenderTFMinoshroom extends BipedRenderer<EntityTFMinoshroom, ModelT
 				BlockState blockstate = Blocks.RED_MUSHROOM.getDefaultState(); // TF: hardcode mushroom state
 				int i = LivingRenderer.getPackedOverlay(entity, 0.0F);
 				ms.push();
-				ms.translate((double)0.2F, (double)-0.35F, 0.5D);
+				ms.translate(0.2F, -0.35F, 0.5D);
 				ms.rotate(Vector3f.YP.rotationDegrees(-48.0F));
 				ms.scale(-1.0F, -1.0F, 1.0F);
 				ms.translate(-0.5D, -0.5D, -0.5D);
 				blockrendererdispatcher.renderBlock(blockstate, ms, buffers, light, i);
 				ms.pop();
 				ms.push();
-				ms.translate((double)0.2F, (double)-0.35F, 0.5D);
+				ms.translate(0.2F, -0.35F, 0.5D);
 				ms.rotate(Vector3f.YP.rotationDegrees(42.0F));
-				ms.translate((double)0.1F, 0.0D, (double)-0.6F);
+				ms.translate(0.1F, 0.0D, -0.6F);
 				ms.rotate(Vector3f.YP.rotationDegrees(-48.0F));
 				ms.scale(-1.0F, -1.0F, 1.0F);
 				ms.translate(-0.5D, -0.5D, -0.5D);
