+++ src/main/java/twilightforest/client/renderer/entity/RenderTFTowerGolem.java
@@ -24,7 +24,7 @@ public class RenderTFTowerGolem<T extends EntityTFTowerGolem, M extends ModelTFT
 	protected void applyRotations(T entity, MatrixStack ms, float ageInTicks, float rotationYaw, float partialTicks) {
 		super.applyRotations(entity, ms, ageInTicks, rotationYaw, partialTicks);
 
-		if (!((double)entity.limbSwingAmount < 0.01D)) {
+		if (!(entity.limbSwingAmount < 0.01D)) {
 			float f = 13.0F;
 			float f1 = entity.limbSwing - entity.limbSwingAmount * (1.0F - partialTicks) + 6.0F;
 			float f2 = (Math.abs(f1 % 13.0F - 6.5F) - 3.25F) / 3.25F;
